backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
pydantic[email]>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
sse-starlette>=2.1.0
